Feat/add zero tvl (#16)
* Added zero tvl

* Small fix

* Fixes to Zero TVL

* Reset synchronize
