+++ package.json
@@ -26,8 +26,8 @@
    "author": "Remy",
    "license": "ISC",
    "dependencies": {
-      "@apollo/client": "^3.6.2",
-      "@blobfishkate/sovryncontractswip": "^0.2.11",
+      "@apollo/client": "^3.6.9",
+      "@blobfishkate/sovryncontractswip": "^0.2.15",
       "apollo-cache-inmemory": "^1.6.6",
       "apollo-client": "^2.6.10",
       "apollo-link-http": "^1.5.17",
@@ -38,15 +38,15 @@
       "cross-fetch": "^3.1.5",
       "express": "^4.18.1",
       "express-request-id": "^1.4.1",
-      "express-validator": "^6.14.1",
+      "express-validator": "^6.14.2",
       "graphql": "^16.5.0",
       "graphql-tag": "^2.12.6",
       "helmet": "^4.6.0",
       "lodash": "^4.17.21",
-      "mathjs": "^10.5.2",
-      "moment": "^2.29.3",
-      "node-fetch": "^3.2.4",
-      "npm": "^8.9.0",
+      "mathjs": "^10.6.4",
+      "moment": "^2.29.4",
+      "node-fetch": "^3.2.10",
+      "npm": "^8.15.1",
       "pg": "^8.7.3",
       "pino": "^6.14.0",
       "pino-http": "^5.8.0",
@@ -56,20 +56,20 @@
       "response-time": "^2.3.2",
       "typeorm": "^0.2.45",
       "uuid": "^8.3.1",
-      "web3": "^1.7.3",
-      "web3-utils": "^1.7.3"
+      "web3": "^1.7.5",
+      "web3-utils": "^1.7.5"
    },
    "devDependencies": {
-      "@babel/cli": "^7.17.10",
-      "@babel/core": "^7.17.10",
-      "@babel/node": "^7.17.10",
-      "@babel/preset-env": "^7.17.10",
+      "@babel/cli": "^7.18.10",
+      "@babel/core": "^7.18.10",
+      "@babel/node": "^7.18.10",
+      "@babel/preset-env": "^7.18.10",
       "@graphql-codegen/cli": "2.6.2",
       "@graphql-codegen/introspection": "2.1.1",
       "@graphql-codegen/typescript": "2.4.8",
       "@graphql-codegen/typescript-document-nodes": "2.2.8",
       "@graphql-codegen/typescript-resolvers": "2.6.1",
-      "@tsconfig/node12": "^1.0.9",
+      "@tsconfig/node12": "^1.0.11",
       "@types/bunyan": "^1.8.8",
       "@types/convict": "^6.1.1",
       "@types/cors": "^2.8.12",
@@ -77,22 +77,22 @@
       "@types/express": "^4.17.13",
       "@types/express-request-id": "^1.4.3",
       "@types/graphql": "^14.5.0",
-      "@types/jest": "^27.5.0",
+      "@types/jest": "^27.5.2",
       "@types/lodash": "^4.14.182",
       "@types/node": "^8.10.66",
       "@types/pino": "^6.3.12",
       "@types/pino-http": "^5.8.1",
       "@types/response-time": "^2.3.5",
       "@types/supertest": "^2.0.12",
-      "apollo": "^2.33.11",
+      "apollo": "^2.34.0",
       "husky": "^7.0.4",
       "jest": "^27.5.1",
-      "nodemon": "^2.0.16",
-      "supertest": "^6.2.3",
-      "ts-jest": "^27.1.4",
-      "ts-node": "^10.7.0",
+      "nodemon": "^2.0.19",
+      "supertest": "^6.2.4",
+      "ts-jest": "^27.1.5",
+      "ts-node": "^10.9.1",
       "ts-standard": "^10.0.0",
-      "typechain": "^8.0.0",
-      "typescript": "^4.6.3"
+      "typechain": "^8.1.0",
+      "typescript": "^4.7.4"
    }
 }
